fix: fixed profile state (store)
