+++ packages/client/src/app/redux/store.ts
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import authSlice from '../../pages/Auth/redux/authSlice'
 import forumSlice from '../../pages/Forum/redux/forumSlice'
-import profileSlice from '../../pages/Profile/redux/profileSlice'
 import leaderboardSlice from '../../pages/Leaderboard/redux/leaderboardSlice'
 import snackSlice from '../../components/Snackbar/redux/snackbarSlice'
 
@@ -10,7 +9,6 @@ export const createStore = (preloadedState: any) => {
     reducer: {
       auth: authSlice,
       forum: forumSlice,
-      profile: profileSlice,
       snack: snackSlice,
       leaderboard: leaderboardSlice,
     },
